Agentic AI Assignment/IN226018202_GenAI/Assignement_Task_3_AI Resume Screening System with Tracing/requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.1
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.9
groq>=0.9.0
pypdf>=4.2.0
langsmith>=0.1.0
